requirements.txt
streamlit>=1.32.0
dashscope>=1.14.0
jieba>=0.42.1
requests>=2.31.0
pillow>=10.2.0
python-dotenv>=1.0.0